Add collision detection to WarpPipe
The commit adds overlap functions OnCollisionBeginOverlap and OnCollisionEndOverlap to detect and manage entity collisions with WarpPipe objects. The interaction with the WarpPipe (MovePlayerUp and MovePlayerDown) is now contingent on these states, improving character movement behavior. A variable bIsGateOpen is introduced to keep track of the open/closed state of the pipe.