Add spawn location functionality to Horace character
In this commit, a new feature is added to the Horace character, allowing to set the player's spawn location. A new protected property, SpawnTransform, has been added to the GameModeBase, along with a public setter function. In the Horace character class, a new method has been added, SetTransformAsNextSpawnLocation, calling this setter. The method is called in the BeginPlay function, setting the spawn location to the character's initial position.