Implement checkpoint functionality and refactor function names in library
In this commit, a checkpoint functionality has been added to the game's checkpoint system. The functionality includes the tracking of player's respawn location, rotation of checkpoint flag and triggering effects when a checkpoint is crossed. Simultaneously, the function naming convention in the utility library has been refactored for better comprehension, changing 'PlayBlockSoundFx' and 'PlayBlockInteractFx' to 'PlaySoundFx' and 'PlayInteractFx', respectively. Minor changes to effect spawning and sound effect triggering in different classes have also been made accordingly.