Implement level timer and audio playback on respawn
The commit includes the implementation of a level timer and audio playback on player respawn. This enhances the user's game experience by adding an extra level of interaction. It also set the default value of TimeLimitMax to 300 seconds. New methods 'SetAudioToPlay', 'StartLevelTimer', and 'RunLevelTimer' were introduced and integrated into the respawn logic.