Refactor game loop and improve timer logic
Several improvements are made to the game logic in this commit. There is a restructuring of the game loop in HA_LevelSettings with the start of the loop now handled by a new method, StartGameLoop(). Enhancements are also made to the timer feature in HA_EndOfLevelFlag and HA_LevelSettings to better manage the time dependencies within the game. These changes result in a cleaner and more efficient game loop and timer management.