Implement main menu UI and functionality
Added new main menu UI with button interactions for play button and animations for the main menu screen. The UI consists of a button play that opens the first level of the game, and two types of animations, an intro and a scaling effect on the button play. Updated GameMapsSettings in DefaultEngine.ini to point to the newly created MainMenu map.