Refactor EndOfLevelFlag class in HoraceAdventure game
Updated the `AHA_EndOfLevelFlag` class. The MoveTime variable was made a `constexpr` for optimization and the check for firework show triggering has been refactored. We no longer check if the last digit of time remaining equals "Fireworks Show #", instead, we directly compare `FireworkShowSpecialNumber` to `HoraceController->GetLevelTime()`. Also, the `FireworkShowSpecialNumber` has been updated to 50.